Add files via upload
The goal of this project was to design and develop an employee management system for the Employee management system. This system was developed to fill existing gaps in the electronic management of employees. The scope of the project included researching and evaluating various Web technologies and analysis and design of employee profiles, electronic time sheets, employee performance reviews, and an integrated hiring process. Development work included construction of the core system framework, creation and management of accounts, employee profiles, and electronic time sheets, and a foundation for future modules.